HW02 3-5

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,46 +1,73 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
-
 import java.util.Collection;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
+    private static final Logger log = LoggerFactory.getLogger(InMemoryMealRepository.class);
+
     private final Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private final AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.meals.forEach(this::save);
-    }
-
-    @Override
-    public Meal save(Meal meal) {
-        if (meal.isNew()) {
-            meal.setId(counter.incrementAndGet());
-            repository.put(meal.getId(), meal);
-            return meal;
+        for (Meal meal : MealsUtil.meals) {
+            repository.put(counter.incrementAndGet(), meal);
         }
-        // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        counter.set(repository.size() + 1);
     }
 
     @Override
-    public boolean delete(int id) {
-        return repository.remove(id) != null;
+    public Meal save(int userId, Meal meal) {
+        log.info("save {}", meal);
+        if (meal.getUserId() == userId) {
+            if (meal.isNew()) {
+                meal.setId(counter.incrementAndGet());
+                repository.put(meal.getId(), meal);
+                return meal;
+            } else
+                // handle case: update, but not present in storage
+                return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        } else {
+            return null;
+        }
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public boolean delete(int userId, int id) {
+        log.info("delete {}", id);
+        if (get(userId, id) != null) {
+            return repository.remove(id) != null;
+        } else {
+            return false;
+        }
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Meal get(int userId, int id) {
+        log.info("get {}", id);
+        if (repository.get(id) != null && repository.get(id).getUserId() == userId) {
+            return repository.get(id);
+        } else {
+            return null;
+        }
+    }
+
+    @Override
+    public Collection<Meal> getAll(int userId) {
+        log.info("getAll");
+        List<Meal> allMeals = repository.values().stream().filter(meal -> meal.getUserId() == userId).collect(Collectors.toList());
+        return allMeals.size() != 0 ? allMeals : null;
     }
 }
